Change the counter type of base::RefCounted from size_t to int
All other reference counts uses int instead of size_t, and as the size
of reference count matters to Blink, base::RefCounted should use 32 bit
ref count to unify to Blink's WTF::RefCounted.

Bug:

+++ base/memory/ref_counted.h
@@ -136,7 +136,7 @@ class BASE_EXPORT RefCountedBase {
   bool CalledOnValidSequence() const;
 #endif
 
-  mutable size_t ref_count_ = 0;
+  mutable int ref_count_ = 0;
 
 #if DCHECK_IS_ON()
   mutable bool needs_adopt_ref_ = false;
